Retrieves the state of a single pin on the GPIO

Retrieves the state of a single pin on the GPIO

Usage

gpioget(gpio_chip = 0, gpio_active = TRUE, gpio_line, gpio_bias = "as-is")

Arguments

gpio_chip

Retrieve the pin info from this chip number. Defaults to chip 0. Chips are listed by gpiodetect()

gpio_active

Set the line active state, or what value to return when signal is high. gpio_active = FALSE inverts the active state of the line

gpio_line

Retrieve state from this pin. Uses BCM numbering. This can be a single number or a vector or line numbers

gpio_bias

Set the line bias to one of the following: as-is (the default), disable, pull-down, pull-up

Value

A vector of 1 or 0 the same length of gpio_line indicating the state of each pin

Examples

# read the value of a single gpio line
gpioget(gpio_chip = 0, gpio_line = 21)
#> Error in system(gpio_sysCall, intern = TRUE): error in running command

# Read two values at the same time.
gpioget(gpio_chip = 0, gpio_line = c(21,19))
#> Error in system(gpio_sysCall, intern = TRUE): error in running command